chore: move SearchJob to separate file
